This changes the GPIO's for the model with the MIC
I haven't done anything with the MIC yet, but to get the rest of the board working I had to change the GPIO's in the YAML file to the way this board has them.  I figured this would help anyone else who might be new and struggling to understand why it isn't working.  I have a few ideas for the MIC such as leaving a VM for someone at the home or speaking the person's name and trigingering HA to alert them that they have a visitor.  If anyone would like to help me with that part that would be amazing.  As always have fun first.